refractored a bit, getting response back  from response and success action with payload, unable to render data to app due to state being undefined, sure there is a mistake some where with the passing of data
