+++ smurfs/src/reducers/add.js
@@ -1,5 +1,5 @@
 import { ADD_SMURF, UPDATE_SMURF } from '../actions/add';
 
 const initialState = {
-  
+  smurf: []
 }
